+++ src/main/kotlin/dev/turingcomplete/intellijgradleutilitiesplugin/gradleenvironment/CollectGradleEnvironmentInformationAction.kt
@@ -6,7 +6,6 @@ import com.intellij.openapi.diagnostic.Logger
 import com.intellij.openapi.progress.ProgressIndicator
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.project.guessProjectDir
-import com.intellij.openapi.util.PropertiesUtil
 import com.intellij.openapi.vfs.VirtualFile
 import com.intellij.ui.ScrollPaneFactory
 import com.intellij.util.ui.JBUI
@@ -18,9 +17,6 @@ import dev.turingcomplete.intellijgradleutilitiesplugin.common.ui.GradleUtilityD
 import org.apache.commons.codec.digest.DigestUtils
 import java.awt.Dimension
 import java.io.FileInputStream
-import java.io.FileReader
-import java.nio.file.Files
-import java.nio.file.Path
 import javax.swing.ScrollPaneConstants
 
 class CollectGradleEnvironmentInformationAction
@@ -115,11 +111,14 @@ class CollectGradleEnvironmentInformationAction
         determineGradleWrapperVersion(projectDir, gradlewExecutable, progressIndicator)
       }
 
-      val gradleWrapperProperties = GradleUtils.findGradleWrapperProperties(projectDir)?.let { gradleWrapperProperties ->
-        readProperties("Reading Gradle wrapper properties...", gradleWrapperProperties.toNioPath(), progressIndicator, handleError)
-      } ?: listOf()
+      val gradleWrapperProperties = readGradleWrapperProperties(projectDir, progressIndicator, handleError)
 
-      return GradleEnvironment.GradleWrapper(gradleWrapperVersion, gradleWrapperJarChecksum, gradleWrapperProperties)
+      val checksumVerificationConfigured = GradleUtils.isChecksumVerificationConfigured(gradleWrapperProperties)
+
+      return GradleEnvironment.GradleWrapper(gradleWrapperVersion,
+                                             gradleWrapperJarChecksum,
+                                             gradleWrapperProperties,
+                                             checksumVerificationConfigured)
     }
     catch (e: Exception) {
       handleError("Failed to collect Gradle wrapper information.", e)
@@ -131,30 +130,6 @@ class CollectGradleEnvironmentInformationAction
     progressIndicator.checkCanceled()
     progressIndicator.text = "Calculating Gradle wrapper JAR checksum..."
     return FileInputStream(gradleWrapperJar.toNioPath().toFile()).use { DigestUtils.sha256Hex(it) }
-  }
-
-  private fun readProperties(progressText: String,
-                             propertiesFile: Path,
-                             progressIndicator: ProgressIndicator,
-                             handleError: (String, Exception) -> Unit): List<Pair<String, String>> {
-
-    progressIndicator.checkCanceled()
-
-    try {
-      progressIndicator.text = progressText
-
-      if (!Files.exists(propertiesFile)) {
-        return listOf()
-      }
-
-      return FileReader(propertiesFile.toFile()).use {
-        PropertiesUtil.loadProperties(it).map { property -> Pair(property.key, property.value) }.sortedBy { it.first }
-      }
-    }
-    catch (e: Exception) {
-      handleError("Failed to read properties file: $propertiesFile", e)
-      return listOf()
-    }
   }
 
   // -- Inner Type -------------------------------------------------------------------------------------------------- //